+++ spec/unit/abstract_mapper/settings_spec.rb
@@ -4,7 +4,7 @@ class AbstractMapper
 
   describe AbstractMapper::Settings do
 
-    let!(:rule) { Test::Rule = Class.new(SoleRule) }
+    let!(:rule) { Test::Rule = Class.new(Rules::Sole) }
     let!(:node) { Test::Node = Class.new(Node) { attribute :foo } }
 
     let(:settings) do